WebApp/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==2.1
dnspython==2.1.0
idna==2.10
paho-mqtt==1.5.1
pymongo==3.11.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
